+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+### Added
+- `earthly ls [<project-ref>]` to list Earthfile targets
+
 ## v0.6.2 - 2021-12-01
 
 ### Fixed